requirements.txt
python-Levenshtein
fuzzywuzzy
sentence-transformers
torch
pandas
tqdm
scikit-learn